change sequence to int64

+++ common/data_def/device_data.hpp
@@ -13,7 +13,7 @@ struct DeviceRawData {
     int32_t length;
     int16_t device_type;
     int16_t device_data_type;
-    int32_t sequence;
+    int64_t sequence;
     int64_t timestamp_receive_ns;
     uint8_t rawdata_buf[0];
 };
@@ -22,7 +22,7 @@ struct SensorData {
     int32_t length;
     int16_t device_type;
     int16_t device_data_type;
-    int32_t sequence;
+    int64_t sequence;
     int64_t timestamp_receive_ns;
     int64_t timestamp_intrinsic_ns;
     uint8_t data_buf[0];
